fix email theming issue
