+++ packages/core/src/controllers/ThemeController.ts
@@ -1,5 +1,6 @@
 import { proxy, subscribe as sub, snapshot } from 'valtio/vanilla'
 import type { ThemeMode, ThemeVariables } from '../utils/TypeUtil.js'
+import { ConnectorController } from './ConnectorController.js'
 
 // -- Types --------------------------------------------- //
 export interface ThemeControllerState {
@@ -23,10 +24,33 @@ export const ThemeController = {
 
   setThemeMode(themeMode: ThemeControllerState['themeMode']) {
     state.themeMode = themeMode
+    try {
+      const emailConnector = ConnectorController.getEmailConnector()
+
+      if (emailConnector) {
+        emailConnector.provider.syncTheme({
+          themeMode: ThemeController.getSnapshot().themeMode
+        })
+      }
+    } catch {
+      // eslint-disable-next-line no-console
+      console.info('Unable to sync theme to email connector')
+    }
   },
 
   setThemeVariables(themeVariables: ThemeControllerState['themeVariables']) {
     state.themeVariables = { ...state.themeVariables, ...themeVariables }
+    try {
+      const emailConnector = ConnectorController.getEmailConnector()
+      if (emailConnector) {
+        emailConnector.provider.syncTheme({
+          themeVariables: ThemeController.getSnapshot().themeVariables
+        })
+      }
+    } catch {
+      // eslint-disable-next-line no-console
+      console.info('Unable to sync theme to email connector')
+    }
   },
 
   getSnapshot() {
